requirements.txt
pillow
requests
paho-mqtt
sqlalchemy
matplotlib
numpy